cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adp-cli"
version = "0.1.0"
description = "Agent Definition Protocol CLI"
authors = [{name = "ADP Maintainers"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "typer>=0.12",
  "jsonschema>=4.21",
  "rich>=13.0",
  "zipfile36; python_version < '3.11'"
]

[project.optional-dependencies]
test = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "coverage[toml]>=7.5"
]

[project.scripts]
adp = "adp_cli.main:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
branch = true
source = ["adp_cli"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 0

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
where = ["."]
include = ["adp_cli*"]

[tool.typer]
completion = true
